Split ResultlistContainer to abstract NowProvider
